requirements.txt
Flask==0.10.1
pycryptodome==3.4.3
flask-cors==3.0.8
psutil==5.7.0